Define "homeworld" level type
This is a basic level type that simply identifies homeworlds. No other
information is required that is not already present in other levels.

The "kind" property has been added to levels to help distinguish types
instead of just on properties.

The levels list at the root now looks for specific kinds of levels
instead of general level types.