Generate angular project
This starting project was generated using
`ng new -g --strict --style sass --routing` in the folder `app/`. The
only change in the configuration was to make the output directory
`/dist/app` from the root of the repository. This is to localise
different versions of the application (e.g. a raw text document vs a
web app).