Add GitHub Action to build angular app
This is a simple job that navigates to the `app/` directory and runs the
equivalent of `ng build` to deposit it into `dist/app`. Further steps
are required to create the required pull requests.

Tests have been commented out due to errors requiring the Google Chrome
browser. This is needed for the Karma tests, but an appropriate binary
(or its dependencies) cannot be found for it yet.