Transform service to use Observables
The Spyro2Service abstract class has been changed to expect Obseravbles
of levels and locales rather than raw data structures. This is to enable
the code structure to handle asynchronous calls from dependencies (e.g.
HTTP calls from the HttpClient).

A lot of rework was required for the implementation to handle this. New
functionality was added to handle stub databases and transform them via
the Observer methods.

Dummy data was extended to more levels to further test that these
Observables work as intended.