Update warn and error sizes for styles
Due to the component styles using bootstrap, they end up being too large
when precompiled in their current form. I believe this is because even
when a global style is set for Sass, the context it provides is not
inherited for dependent Sass files.

I would like to address this at some point and revert the styles size
restrictions back to what they were, but it may require some thought on
what is global and how component styles are specified.